requirements.txt
requests
cryptography
mutagen
click
pyinquirer
colorama
rich
autopep8
pylint
deezer-py